+++ src/benchq/compilation/gate_stitching.py
@@ -1,4 +1,6 @@
-import time
+################################################################################
+# © Copyright 2022-2023 Zapata Computing Inc.
+################################################################################
 from typing import List, Set, Tuple
 
 import networkx as nx
@@ -8,7 +10,11 @@ from .graph_sim_mini import get_vertices
 
 
 def get_algorithmic_graph_from_gate_stitching(circuit: Circuit) -> nx.Graph:
-    """Get the graph corresponding to the state generated by the circuit.
+    """Get the graph corresponding to the state generated by the circuit using
+    a graph state simulator (graph_sim) from the paper "Fast simulation of stabilizer
+    circuits using a graph state representation" by Simon Anders, Hans J. Briegel. We
+    convert the circuit to the ICM form and then use the graph state simulator
+    to directly get the graph corresponding to the state generated by the circuit.
 
     Args:
         circuit (Circuit): the circuit to convert to a graph
@@ -16,23 +22,9 @@ def get_algorithmic_graph_from_gate_stitching(circuit: Circuit) -> nx.Graph:
     Returns:
         nx.Graph: graph corresponding to the state generated by the circuit
     """
-    print("converting circuit to icm form")
-    start = time.time()
     icm_circuit = get_icm(circuit)
-    end = time.time()
-    print(f"Converting to icm took {end - start} seconds")
-
-    print("getting graph vertices")
-    start = time.time()
     vertices = get_vertices(icm_circuit)
-    end = time.time()
-    print(f"getting graph vertices took {end - start} seconds")
-
-    print("Convert vertices to networkx graph")
-    start = time.time()
-    algorithmic_graph = get_networkx(vertices)
-    end = time.time()
-    print(f"Converting vertices to networkx graph took {end - start} seconds")
+    algorithmic_graph = get_networkx_graph_from_vertices(vertices)
 
     return algorithmic_graph
 
@@ -71,13 +63,12 @@ def get_icm(circuit: Circuit, gates_to_decompose=["T", "T_Dagger"]) -> Circuit:
     return Circuit(icm_circuit)
 
 
-def get_networkx(vertices: List[Tuple[int, Set[int]]]) -> nx.Graph:
-    """Convert the vertices of the graph to a networkx graph.
+def get_networkx_graph_from_vertices(vertices: List[Tuple[int, Set[int]]]) -> nx.Graph:
+    """Convert the vertices of a graph to a networkx graph.
 
     Args:
-        vertices (list): list of tuples of the form
-            (local clifford operation, adjacency list) corresponding to each node
-            in the graph
+        vertices (list(tuple(int, set(int)))): list of vertices of the graph state
+            of the form (lco, neighbors).
 
         Returns:
             nx.Graph: the networkx graph with those vertices
